moved file to proper folder + changed wrong namespaces
