+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Analytics/DCL.Analytics.asmdef
@@ -11,8 +11,7 @@
     "GUID:e0eedfa2deb9406daf86fd8368728e39",
     "GUID:fa7b3fdbb04d67549916da7bd2af58ab",
     "GUID:ca4e81cdd6a34d1aa54c32ad41fc5b3b",
-    "GUID:166b65e6dfc848bb9fb075f53c293a38",
-    "GUID:c6e727f7851314e679212856f4ce3e53"
+    "GUID:166b65e6dfc848bb9fb075f53c293a38"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
