+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Analytics/EventBased/ChatAnalytics.cs
@@ -2,7 +2,7 @@
 using Segment.Serialization;
 using System;
 
-namespace DCL.PerformanceAndDiagnostics.Analytics
+namespace DCL.PerformanceAndDiagnostics.Analytics.EventBased
 {
     public class ChatEventsAnalytics : IDisposable
     {
