+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Analytics/EventBased/DCL.Analytics.EventBased.asmdef
@@ -22,7 +22,8 @@
         "GUID:ca926b7b13255b84199c21c5e2d9dfa4",
         "GUID:9ca29a7d75ab84bab806a20bbd350fd8",
         "GUID:f82471a6db95848d88faad2ceb31c9c9",
-        "GUID:887770ff22b647ac864047b3c1af9a8b"
+        "GUID:887770ff22b647ac864047b3c1af9a8b",
+        "GUID:c6e727f7851314e679212856f4ce3e53"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
